# More performance enhancements