--- a/app/scripts/infuse/dimInfuseItem.directive.js
+++ b/app/scripts/infuse/dimInfuseItem.directive.js
@@ -15,6 +15,7 @@
       template: [
         '<div title="{{ vm.item.primStat.value }} {{ vm.item.name }}" alt="{{ vm.item.primStat.value }} {{ vm.item.name }}" class="item" ng-class="{ \'search-hidden\': !vm.item.visible, \'search-item-hidden\': vm.item.visible === false && vm.hideFilteredItems === true, \'complete\': vm.item.complete }">',
         '  <div ng-if="vm.item.locked" class="item-name"><div class="locked"></div></div>',
+        '  <div class="img"></div>',
         '  <div class="damage-type" ng-if="!vm.item.itemStat && vm.item.sort === \'Weapons\'" ng-class="\'damage-\' + vm.item.dmg"></div>',
         '  <div class="item-stat" ng-if="vm.item.primStat.value" ng-class="\'stat-damage-\' + vm.item.dmg">{{ vm.item.primStat.value }}</div>',
         '</div>'
@@ -26,10 +27,12 @@
         var vm = scope.vm;
         $('<img/>').attr('src', 'http://www.bungie.net' + vm.item.icon).load(function() {
            $(this).remove();
-          element[0].style.backgroundImage = 'url(' + 'http://www.bungie.net' + vm.item.icon + ')';
+          element[0].querySelector('.img')
+            .style.backgroundImage = 'url(' + 'http://www.bungie.net' + vm.item.icon + ')';
         }).error(function() {
            $(this).remove();
-          element[0].style.backgroundImage = 'url(' + chrome.extension.getURL(vm.item.icon) + ')';
+          element[0].querySelector('.img')
+            .style.backgroundImage = 'url(' + chrome.extension.getURL(vm.item.icon) + ')';
         });
       }
     };