--- a/app/scripts/store/dimStoreHeading.directive.js
+++ b/app/scripts/store/dimStoreHeading.directive.js
@@ -18,50 +18,26 @@
       },
       link: Link,
       template: [
-        '<div class="character-box" ng-class="vm.isGuardian ? \'\' : \'vault-box\'">',
-        '  <div class="emblem" ng-show="vm.isGuardian"></div>',
-        '  <div class="class">{{ vm.class || "Vault" }}</div>',
-        '  <div class="race-gender" ng-show="vm.isGuardian">{{ vm.race }} {{ vm.gender }}</div>',
-        '  <div class="level" ng-show="vm.isGuardian" ng-class="vm.isPrestigeLevel ? \'prestige\' : \'\'">Level {{ vm.level }}</div>',
-        '  <div class="level powerLevel" ng-show="vm.isGuardian" ng-class="vm.isPrestigeLevel ? \'prestige\' : \'\'">{{ vm.store.powerLevel }}</div>',
-        '  <div class="glimmer" ng-show="!vm.isGuardian"> {{ vm.store.glimmer }} <img src="/images/glimmer.png"></div>',
-        '  <div class="legendaryMarks" ng-show="!vm.isGuardian"> {{ vm.store.legendaryMarks }} <img src="/images/legendaryMarks.png"></div>',
-        '  <div class="levelBar" ng-show="vm.isGuardian">',
-        '    <div class="barFill" ng-class="vm.isPrestigeLevel ? \'prestige\' : \'\'" ng-style="{width: vm.percentToNextLevel + \'%\'}"></div>',
+        '<div class="character-box" ng-class="::{ \'vault-box\': !vm.isGuardian }" ng-style="{ \'background-image\': \'url(http://bungie.net\' + vm.store.background + \')\' }">',
+        '  <div class="emblem" ng-if="::vm.isGuardian" ng-style="{ \'background-image\': \'url(http://bungie.net\' + vm.store.icon + \')\' }"></div>',
+        '  <div class="class">{{:: vm.store.class || "Vault" }}</div>',
+        '  <div class="race-gender" ng-if="::vm.isGuardian">{{:: vm.store.race }} {{:: vm.store.gender }}</div>',
+        '  <div class="level" ng-if="::vm.isGuardian" ng-class="{ prestige: vm.isPrestigeLevel, maxLevel: (vm.store.level >= 40) }">Level {{ vm.store.level }}</div>',
+        '  <div class="level powerLevel" ng-if="vm.isGuardian" ng-class="{ prestige: vm.isPrestigeLevel }">{{ vm.store.powerLevel }}</div>',
+        '  <div class="glimmer" ng-if="::!vm.isGuardian"> {{ vm.store.glimmer }} <img src="/images/glimmer.png"></div>',
+        '  <div class="legendaryMarks" ng-if="::!vm.isGuardian"> {{ vm.store.legendaryMarks }} <img src="/images/legendaryMarks.png"></div>',
+        '  <div class="levelBar" ng-if="::vm.isGuardian">',
+        '    <div class="barFill" ng-class="{ prestige: vm.isPrestigeLevel }" ng-style="{width: vm.store.percentToNextLevel + \'%\'}"></div>',
         '  </div>',
         '</div>',
-        '<div class="loadout-button" ng-show="vm.isGuardian" ng-click="vm.openLoadoutPopup($event)"><i class="fa fa-chevron-down"></i></div>',
-        '<div loadout-id="{{ vm.store.id }}" style="position: relative; top: 50px;"></div>'
+        '<div class="loadout-button" ng-if="::vm.isGuardian" ng-click="vm.openLoadoutPopup($event)"><i class="fa fa-chevron-down"></i></div>',
+        '<div loadout-id="{{:: vm.store.id }}" style="position: relative; top: 50px;"></div>'
       ].join('')
     };
 
     function Link(scope, element) {
       var vm = scope.vm;
       var dialogResult = null;
-
-      scope.$watch('vm.store.percentToNextLevel', function(percentToNextLevel) {
-        vm.percentToNextLevel = percentToNextLevel;
-      });
-
-      scope.$watch('vm.store.level', function(level) {
-        vm.level = level;
-      });
-
-      scope.$watch('vm.store.background', function(background) {
-        vm.characterBoxUrl = 'http://bungie.net' + background;
-
-        if (vm.isGuardian) {
-          element[0].querySelector('.character-box').style.backgroundImage = 'url(' + vm.characterBoxUrl + ')';
-        }
-      });
-
-      scope.$watch('vm.store.icon', function(icon) {
-        vm.emblemUrl = 'http://bungie.net' + icon;
-
-        if (vm.isGuardian) {
-          element[0].querySelector('.emblem').style.backgroundImage = 'url(' + vm.emblemUrl + ')';
-        }
-      });
 
       $(document).ready(function() {
         element.scrollToFixed({
@@ -106,20 +82,7 @@
           dialogResult.close();
         }
       };
-
       element.addClass('character');
-
-      if (vm.isGuardian) {
-        element[0].querySelector('.character-box')
-          .style.backgroundImage = 'url(' + vm.characterBoxUrl + ')';
-        element[0].querySelector('.emblem')
-          .style.backgroundImage = 'url(' + vm.emblemUrl + ')';
-
-        if (vm.maxLevel) {
-          element[0].querySelector('.level')
-            .classList.add('maxLevel');
-        }
-      }
     }
   }
 
@@ -127,16 +90,6 @@
 
   function StoreHeadingCtrl($scope) {
     var vm = this;
-
     vm.isGuardian = (vm.store.id !== 'vault');
-    vm.class = vm.store.class;
-    vm.level = vm.store.level;
-    vm.race = vm.store.race;
-    vm.gender = vm.store.gender;
-    vm.isPrestigeLevel = vm.store.isPrestigeLevel;
-    vm.percentToNextLevel = vm.store.percentToNextLevel;
-    vm.maxLevel = (vm.store.level >= 20);
-    vm.characterBoxUrl = 'http://bungie.net' + vm.store.background;
-    vm.emblemUrl = 'http://bungie.net' + vm.store.icon;
   }
 })();