--- a/app/styles/main.css
+++ b/app/styles/main.css
@@ -688,7 +688,7 @@ img.void {
   position: absolute;
   z-index: 2;
 }
-.item {
+.item > .img {
   width: 44px;
   height: 44px;
   background-size: cover;