I already pushed a few improvements to `dev` (accidentally, sorry):

* https://github.com/DestinyItemManager/DIM/commit/e8a4db0e59bc7c025d6efc8a6020f5afee970b62
* https://github.com/DestinyItemManager/DIM/commit/264f600d52786754355429dbcf1c328123d0b83c
* https://github.com/DestinyItemManager/DIM/commit/febc9520857931f243dc62825f99d1addba133e6
* https://github.com/DestinyItemManager/DIM/commit/6ec5001ea7114b2c4e64c7ab5a01951049e8be5b
* https://github.com/DestinyItemManager/DIM/commit/396a809cd793752d32eb909961f31d395f665b8f
* https://github.com/DestinyItemManager/DIM/commit/db4bb76df7a5774a91afd1ad8ba180a4c91b313d
* https://github.com/DestinyItemManager/DIM/commit/6df6644a5096323c6d24d2dc4190a680065474b6

Plus what's in this PR. After these changes, on my machine, startup goes from ~5s to ~3s, and average digest time is down to ~15ms from ~40ms. Everything feels a lot snappier. I'll play with it a bit on my side before recommending a release.